Professional Sample: Policy and Workflow Creation
It was found at this particular university in China that there was no policy for dealing with academic integrity. I was charged with drafting not only a policy but working procedures for documenting incidents and handling appeals. This was, to my knowledge, largely adopted, although much of the discussion occurred at administrative levels to which I was not privy.